requirements.txt
django-sortedm2m~=3.1.1
django-reversion~=5.0.4
django-taggit~=4.0.0
netjsonconfig @ https://github.com/edge-servers/netjsonconfig/tarball/master
django-x509 @ https://github.com/edge-servers/django-x509/tarball/master
django-loci @ https://github.com/edge-servers/django-loci/tarball/master
django-flat-json-widget @ https://github.com/edge-servers/django-flat-json-widget/tarball/master
immunity-users @ https://github.com/edge-servers/immunity-users/tarball/master
immunity-utils[celery] @ https://github.com/edge-servers/immunity-utils/tarball/master
immunity-notifications @ https://github.com/edge-servers/immunity-notifications/tarball/master
immunity-ipam @ https://github.com/edge-servers/immunity-ipam/tarball/master
djangorestframework-gis~=0.18.0
paramiko[ed25519]~=3.2.0
scp~=0.14.2
django-cache-memoize~=0.1.0
shortuuid~=1.0.1
netaddr~=0.8.0
django-import-export~=3.3.0
